refactor: CLIに表示する部分の処理をモジュール分割
